+++ src/path.rs
@@ -9,6 +9,66 @@ Exp    Trm    Crs-Sec    Course Name    Teacher    Room    Enroll    Leave
 7(A-D)    S1    MAT473-2    Linear Algebra    Brummet, Evan    A135    08/19/2024    01/19/2025
 8(A-D)    S1    MAT474-1    Abstract Algebra    Fogel, Micah    A155    08/19/2024    01/19/2025
 */
+
+use regex::Regex;
+
+use crate::{name_to_id, pathfinding, Node};
+struct ScheduleInfo {
+    mods: Vec<String>,
+    semester: Vec<String>,
+    short_name: Vec<String>,
+    long_name: Vec<String>,
+    teacher: Vec<String>,
+    room: Vec<String>,
+    start: Vec<String>,
+    end: Vec<String>,
+}
+
+#[derive(Debug)]
+enum Semester {
+    S1,
+    S2,
+    Year,
+}
+
+#[derive(Debug, PartialEq, Eq, Hash)]
+enum Day {
+    A,
+    B,
+    C,
+    D,
+    I,
+}
+
+#[derive(Debug)]
+pub struct Class {
+    days: Vec<Day>,
+    mods: Vec<u8>,
+    semester: Semester,
+    short_name: String,
+    long_name: String,
+    teacher: String,
+    room: String,
+    start: String,
+    end: String,
+}
+
+#[derive(Debug)]
+pub struct DailyNode {
+    pub anode: Option<Vec<Vec<usize>>>,
+    pub bnode: Option<Vec<Vec<usize>>>,
+    pub inode: Option<Vec<Vec<usize>>>,
+    pub cnode: Option<Vec<Vec<usize>>>,
+    pub dnode: Option<Vec<Vec<usize>>>,
+}
+
+pub enum EnterExit {
+    WestMain,
+    EastMain,
+    D13,
+    D6,
+}
+
 fn split_semesters(input: &str) -> (String, String) {
     let mut sem1 = String::new();
     let mut sem2 = String::new();
@@ -107,8 +167,6 @@ fn resolve_semester(input: &str) -> Result<Vec<Class>, String> {
     })
 }
 
-use regex::Regex;
-
 fn parse_day(day_str: &str) -> Result<Vec<Day>, String> {
     let re = Regex::new(r"^[ ABCDI,-]*$").unwrap();
     if !re.is_match(day_str) {
@@ -128,7 +186,14 @@ fn parse_day(day_str: &str) -> Result<Vec<Day>, String> {
                     let start = start_char.chars().next().unwrap();
                     let end = rest.chars().next().unwrap_or(start);
                     for d in start..=end {
-                        if let Some(result) = parse_day_char(d) {
+                        if let Some(result) = match d {
+                            'A' => Some(Day::A),
+                            'B' => Some(Day::B),
+                            'C' => Some(Day::C),
+                            'D' => Some(Day::D),
+                            'I' => Some(Day::I),
+                            _ => None,
+                        } {
                             days.push(result);
                         } else {
                             return Err(format!("Unknown day pattern: {day}"));
@@ -141,17 +206,6 @@ fn parse_day(day_str: &str) -> Result<Vec<Day>, String> {
         }
     }
     Ok(days)
-}
-
-fn parse_day_char(ch: char) -> Option<Day> {
-    match ch {
-        'A' => Some(Day::A),
-        'B' => Some(Day::B),
-        'C' => Some(Day::C),
-        'D' => Some(Day::D),
-        'I' => Some(Day::I),
-        _ => None,
-    }
 }
 
 fn parse_mods(mod_str: &str) -> Result<Vec<u8>, String> {
@@ -202,13 +256,14 @@ fn sort_by_day(schedule_info: &ScheduleInfo) -> Result<Vec<Class>, String> {
             };
 
             // Flatten all the parsed days and mods into one structure
-            let (all_days, all_mods): (Vec<_>, Vec<_>) = parsed_mods_days
-                .into_iter()
-                .fold((vec![], vec![]), |(mut days, mut mods), (d, m)| {
+            let (all_days, all_mods): (Vec<_>, Vec<_>) = parsed_mods_days.into_iter().fold(
+                (vec![], vec![]),
+                |(mut days, mut mods), (d, m)| {
                     days.extend(d);
                     mods.extend(m);
                     (days, mods)
-                });
+                },
+            );
 
             Ok(Class {
                 days: all_days,
@@ -217,7 +272,12 @@ fn sort_by_day(schedule_info: &ScheduleInfo) -> Result<Vec<Class>, String> {
                     "S1" => Semester::S1,
                     "S2" => Semester::S2,
                     "Y24-25" => Semester::Year,
-                    _ => return Err(format!("Unknown semester '{}'", schedule_info.semester[item])),
+                    _ => {
+                        return Err(format!(
+                            "Unknown semester '{}'",
+                            schedule_info.semester[item]
+                        ))
+                    }
                 },
                 short_name: schedule_info.short_name[item].clone(),
                 long_name: schedule_info.long_name[item].clone(),
@@ -229,7 +289,6 @@ fn sort_by_day(schedule_info: &ScheduleInfo) -> Result<Vec<Class>, String> {
         })
         .collect()
 }
-
 
 pub fn path(weekly_schedule: &Vec<Class>) -> Result<[[String; 8]; 5], String> {
     let mut weekly_path: [[String; 8]; 5] = [
@@ -276,42 +335,131 @@ pub fn path(weekly_schedule: &Vec<Class>) -> Result<[[String; 8]; 5], String> {
 
     Ok(weekly_path)
 }
-struct ScheduleInfo {
-    mods: Vec<String>,
-    semester: Vec<String>,
-    short_name: Vec<String>,
-    long_name: Vec<String>,
-    teacher: Vec<String>,
-    room: Vec<String>,
-    start: Vec<String>,
-    end: Vec<String>,
-}
 
-#[derive(Debug)]
-enum Semester {
-    S1,
-    S2,
-    Year,
-}
+#[allow(clippy::too_many_lines)]
+pub fn node_find_func(
+    schedule: &[[String; 8]; 5],
+    mut nodes: Vec<Node>,
+    entrance: &EnterExit,
+    exit: &EnterExit,
+    checked: bool,
+) -> Result<(DailyNode, Vec<Node>), String> {
+    let mut master_vec: Vec<Vec<Option<[usize; 2]>>> = Vec::new();
 
-#[derive(Debug, PartialEq, Eq, Hash)]
-enum Day {
-    A,
-    B,
-    C,
-    D,
-    I,
-}
+    for day in schedule {
+        let mut vec: Vec<Option<[usize; 2]>> = Vec::new();
 
-#[derive(Debug)]
-pub struct Class {
-    days: Vec<Day>,
-    mods: Vec<u8>,
-    semester: Semester,
-    short_name: String,
-    long_name: String,
-    teacher: String,
-    room: String,
-    start: String,
-    end: String,
+        for (num, mut class) in day.iter().enumerate() {
+            if num == 3 {
+                vec.push(None);
+            }
+            let earlyclass = class.to_lowercase();
+            class = &earlyclass;
+
+            if class.is_empty() {
+                continue;
+            }
+
+            let start_room = name_to_id(class, &nodes)
+                .ok_or(format!("The room '{class}' was not recognized"))?;
+
+            for offset in 1..8 - num {
+                if let Some(next_class) = day.get(num + offset) {
+                    if !next_class.is_empty() {
+                        let next_room = name_to_id(&next_class.to_lowercase(), &nodes)
+                            .ok_or(format!("The room '{next_class}' was not recognized"))?;
+                        if start_room != next_room {
+                            vec.push(Some([start_room, next_room]));
+                        }
+                        break;
+                    }
+                }
+            }
+        }
+        master_vec.push(vec);
+    }
+
+    let mut dailynode = DailyNode {
+        anode: None,
+        bnode: None,
+        inode: None,
+        cnode: None,
+        dnode: None,
+    };
+    for (num, day) in master_vec.into_iter().enumerate() {
+        let mut dayvec: Vec<Vec<usize>> = Vec::new();
+        if day.get(num).is_some() {
+            let shortest_path_st = pathfinding::time_path(
+                {
+                    match entrance {
+                        EnterExit::WestMain => 146,
+                        EnterExit::EastMain => 0,
+                        EnterExit::D13 => 145,
+                        EnterExit::D6 => 147,
+                    }
+                },
+                match day.first().unwrap() {
+                    Some(val) => val[0],
+                    None => 55,
+                },
+                &mut nodes,
+            );
+            dayvec.push(shortest_path_st);
+        }
+
+        for (iter, vecpath) in day.clone().into_iter().enumerate() {
+            if vecpath.is_none() {
+                if checked && day[iter.saturating_sub(1)].is_some() {
+                    let to_lex: Vec<usize> = pathfinding::time_path(
+                        day[iter.saturating_sub(1)].unwrap()[1],
+                        55,
+                        &mut nodes,
+                    );
+                    let from_lex: Vec<usize> =
+                        pathfinding::time_path(55, day[iter + 1].unwrap()[1], &mut nodes);
+                    dayvec.push(to_lex);
+                    dayvec.push(from_lex);
+                } else if day[iter.saturating_sub(1)].is_some() {
+                    let to_norm: Vec<usize> = pathfinding::time_path(
+                        day[iter.saturating_sub(1)].unwrap()[1],
+                        day[iter + 1].unwrap()[1],
+                        &mut nodes,
+                    );
+                    dayvec.push(to_norm);
+                }
+            } else if day[iter.saturating_sub(1)].is_some() {
+                let shortest_path =
+                    pathfinding::time_path(vecpath.unwrap()[0], vecpath.unwrap()[1], &mut nodes);
+                dayvec.push(shortest_path);
+            }
+        }
+        if day.get(num).is_some() {
+            let shortest_path_en = pathfinding::time_path(
+                match day.last().unwrap() {
+                    Some(result) => result[1],
+                    None => 55,
+                },
+                {
+                    match exit {
+                        EnterExit::WestMain => 146,
+                        EnterExit::EastMain => 0,
+                        EnterExit::D13 => 145,
+                        EnterExit::D6 => 147,
+                    }
+                },
+                &mut nodes,
+            );
+            dayvec.push(shortest_path_en);
+        }
+
+        match num {
+            0 => dailynode.anode = Some(dayvec),
+            1 => dailynode.bnode = Some(dayvec),
+            2 => dailynode.inode = Some(dayvec),
+            3 => dailynode.cnode = Some(dayvec),
+            4 => dailynode.dnode = Some(dayvec),
+            _ => return Err(format!("Unexpected num {num}")),
+        }
+    }
+    Ok((dailynode, nodes))
 }
